Layer 파일 추가. Error관련 수정요망

+++ Client/Headers/constant.h
@@ -1,7 +1,7 @@
 #ifndef constant_h__
 #define constant_h__
 
-const DWORD WINCX = 1680;
-const DWORD WINCY = 1050;
+const unsigned long WINCX = 1680;
+const unsigned long WINCY = 1050;
 
 #endif // constant_h__
